pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.1"
description = "hexlet tasks manager"
authors = ["Vadim Berezhnoy  <gpopperony@gmail.com>"]
readme = "README.md"
repository = "https://github.com/popperony/python-project-52"
packages = [{ include = "task_manager" }]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
django = "^5.0.2"
dj-database-url = "^2.1.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"
django-bootstrap5 = "^23.4"
django-filter = "^24.1"
rollbar = "^0.16.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
